+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 1. 【server】修复 在线构建仓库文件存在`.env`文件时带来构建结果异常问题（感谢@蹉跎岁月）
 2. 【agent】优化 项目`清空上传文件`等待文件上传完成后再清空、上传文件优先到临时目录完成后再移动至项目目录（感谢[@tiv-cloud_0](https://gitee.com/tiv-cloud_0) [Gitee issues I649NU](https://gitee.com/dromara/Jpom/issues/I649NU) ）
+3. 【server】修复 docker 创建容器部分场景端口无法正常暴露问题（感谢@李中原）
 
 ### ❌ 不兼容功能